+++ src/main/java/io/github/andrelamego/brValidator/birthdate/InvalidBirthDateException.java
@@ -1,0 +1,7 @@
+package io.github.andrelamego.brValidator.birthdate;
+
+public class InvalidBirthDateException extends RuntimeException {
+    public InvalidBirthDateException(String message) {
+        super(message);
+    }
+}