+++ src/main/java/io/github/andrelamego/brValidator/cep/InvalidCepException.java
@@ -1,0 +1,7 @@
+package io.github.andrelamego.brValidator.cep;
+
+public class InvalidCepException extends RuntimeException {
+    public InvalidCepException(String message) {
+        super(message);
+    }
+}