+++ src/main/java/io/github/andrelamego/brValidator/phone/InvalidPhoneException.java
@@ -1,0 +1,7 @@
+package io.github.andrelamego.brValidator.phone;
+
+public class InvalidPhoneException extends RuntimeException {
+    public InvalidPhoneException(String message) {
+        super(message);
+    }
+}