+++ src/test/java/io/github/andrelamego/brValidator/cnpj/CnpjValidationServiceTests.java
@@ -1,7 +1,5 @@
-package io.github.andrelamego.brValidator.service;
+package io.github.andrelamego.brValidator.cnpj;
 
-import io.github.andrelamego.brValidator.cnpj.CnpjValidationService;
-import io.github.andrelamego.brValidator.cnpj.InvalidCnpjException;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;