+++ src/test/java/io/github/andrelamego/brValidator/cpf/CpfValidationServiceTests.java
@@ -1,6 +1,5 @@
-package io.github.andrelamego.brValidator.service;
+package io.github.andrelamego.brValidator.cpf;
 
-import io.github.andrelamego.brValidator.cpf.CpfValidationService;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;