+++ src/test/java/io/github/andrelamego/brValidator/password/PasswordValidationServiceTests.java
@@ -1,6 +1,5 @@
-package io.github.andrelamego.brValidator.service;
+package io.github.andrelamego.brValidator.password;
 
-import io.github.andrelamego.brValidator.password.PasswordValidationService;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;