+++ src/test/java/io/github/andrelamego/brValidator/password/ValidPasswordIntegrationTest.java
@@ -1,7 +1,5 @@
-package io.github.andrelamego.brValidator.annotation;
+package io.github.andrelamego.brValidator.password;
 
-import io.github.andrelamego.brValidator.password.PasswordValidatorAutoConfiguration;
-import io.github.andrelamego.brValidator.password.ValidPassword;
 import jakarta.validation.Validator;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;